move unload test to init
